DDF-2773: Rebase
